+++ vite.config.ts
@@ -29,7 +29,7 @@ export default defineConfig(({ mode }) => {
                   build: {
                     outDir: "dist-electron",
                     rollupOptions: {
-                      external: ["electron"]
+                      external: ["electron", "discord-rpc", "electron-updater"]
                     }
                   }
                 }